Add "Add to Desktop" and "Add to Panel" actions.
Makes the applet tell the viewer about its containment context
when toggling the viewer and then catch D-Bus signals from the
viewer to implement add requests. The way the applet runs the
viewer has changed a little to simplify the addition of the
above by dropping the --show <screen> argument that was redun-
dant with the toggle() D-Bus method.