Add a fixed copy of PlasmaCore.DialogProxy.
PlasmaCore.DialogProxy.visible's setter contains a workaround
for a situation that isn't applicable to Homerun Kicker and
causes visual artifacts during dialog close there. Due to the
release timeline and the unclear solution, add a copy for now
to have this in the bag until it can be proposed and resolved
upstream.